refactor(dotenv-schema): create new crate
