+++ dotenv-linter/Cargo.toml
@@ -20,13 +20,7 @@ stub_check_version = []
 clap = { version = "4.5", features = ["cargo", "env"] }
 colored = "3.0"
 dotenv-finder = { version = "0.1", path = "../dotenv-finder" }
-email_address = "0.2"
-regex = "1.11"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-serde_regex = "1.1"
-serde_with = "3.14"
-url = "2.5"
+dotenv-schema = { version = "0.1", path = "../dotenv-schema" }
 
 [dependencies.update-informer]
 version = "1.3"
@@ -35,6 +29,8 @@ optional = true
 [dev-dependencies]
 assert_cmd.workspace = true
 criterion = "0.7"
+regex = "1.11"
+serde_json = "1.0"
 tempfile.workspace = true
 
 [target.'cfg(windows)'.dev-dependencies]
