+++ dotenv-schema/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "dotenv-schema"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+include = ["/src"]
+license.workspace = true
+repository.workspace = true
+description = "A crate to validate .env files against schemas"
+
+[dependencies]
+email_address = "0.2"
+regex = "1.11"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+serde_regex = "1.1"
+serde_with = "3.14"
+url = "2.5"
+
+[dev-dependencies]
+tempfile.workspace = true
